disaster-logistics-ai/requirements.txt
# Dataset download
kagglehub>=0.2.0

# Deep learning
torch>=2.0.0
torchvision>=0.15.0

# Image processing
Pillow>=9.0.0
opencv-python>=4.7.0

# Data science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0
